allow option to use a remote git server

+++ repo/repo.go
@@ -2,6 +2,7 @@ package repo
 
 import (
 	"errors"
+	"gemigit/config"
 	"gemigit/db"
 	"io"
 	"os"
@@ -10,27 +11,73 @@ import (
 	"github.com/go-git/go-git/v5/plumbing"
 	"github.com/go-git/go-git/v5/plumbing/object"
 	"github.com/go-git/go-git/v5/plumbing/storer"
+	"github.com/go-git/go-git/v5/storage"
+	"github.com/go-git/go-git/v5/storage/memory"
 )
 
 var rootPath string
+var cache storage.Storer
+var repositories = make(map[string]*git.Repository)
 
 func Init(path string) error {
+	if config.Cfg.Git.Remote.Enabled {
+		cache = memory.NewStorage()
+		return nil
+	}
 	rootPath = path
 	return os.MkdirAll(path, 0700)
 }
 
 func InitRepo(name string, username string) error {
+	if config.Cfg.Git.Remote.Enabled {
+		err := request("api/" + config.Cfg.Git.Remote.Key + "/init/" +
+			       username + "/" + name)
+		return err
+	}
 	_, err := git.PlainInit(rootPath+"/"+username+"/"+name, true)
 	return err
 }
 
 func RemoveRepo(name string, username string) error {
+	if config.Cfg.Git.Remote.Enabled {
+		err := request("api/" + config.Cfg.Git.Remote.Key + "/rm/" +
+			       username + "/" + name)
+		return err
+	}
 	return os.RemoveAll(rootPath + "/" + username + "/" + name)
+}
+
+func getRepo(name string, username string) (*git.Repository, error) {
+	var repo *git.Repository
+	var err error
+	url := username + "/" + name
+	if !config.Cfg.Git.Remote.Enabled {
+		repo, err = git.PlainOpen(rootPath + "/" + url)
+	} else {
+		var exist bool
+		repo, exist = repositories[url]
+		if exist {
+			err = repo.Fetch(&git.FetchOptions{})
+			if err != nil && err != git.NoErrAlreadyUpToDate {
+				return nil, err
+			}
+			return repo, nil
+		}
+		repo, err = git.Clone(cache, nil,
+		&git.CloneOptions{
+			URL: config.Cfg.Git.Remote.Url + "/" + url,
+		})
+		if err != nil {
+			return nil, err
+		}
+		repositories[url] = repo
+	}
+	return repo, err
 }
 
 func GetCommit(name string, username string,
 	       hash plumbing.Hash) (*object.Commit, error) {
-	repo, err := git.PlainOpen(rootPath + "/" + username + "/" + name)
+	repo, err := getRepo(name, username)
 	if err != nil {
 		return nil, err
 	}
@@ -42,7 +89,7 @@ func GetCommit(name string, username string,
 }
 
 func GetCommits(name string, username string) (object.CommitIter, error) {
-	repo, err := git.PlainOpen(rootPath + "/" + username + "/" + name)
+	repo, err := getRepo(name, username)
 	if err != nil {
 		return nil, err
 	}
@@ -58,7 +105,7 @@ func GetCommits(name string, username string) (object.CommitIter, error) {
 }
 
 func GetRefs(name string, username string) (storer.ReferenceIter, error) {
-	repo, err := git.PlainOpen(rootPath + "/" + username + "/" + name)
+	repo, err := getRepo(name, username)
 	if err != nil {
 		return nil, err
 	}
@@ -74,7 +121,7 @@ func GetRefs(name string, username string) (storer.ReferenceIter, error) {
 }
 
 func getTree(name string, username string) (*object.Tree, error) {
-	repo, err := git.PlainOpen(rootPath + "/" + username + "/" + name)
+	repo, err := getRepo(name, username)
 	if err != nil {
 		return nil, err
 	}
@@ -172,6 +219,11 @@ func GetPrivateFile(name string, username string,
 }
 
 func ChangeRepoDir(name string, username string, newname string) error {
+	if config.Cfg.Git.Remote.Enabled {
+		err := request("api/" + config.Cfg.Git.Remote.Key + "/mv/" +
+			       username + "/" + name + "/" + newname)
+		return err
+	}
 	return os.Rename(rootPath + "/" + username + "/" + name,
 			 rootPath + "/" + username + "/" + newname)
 }
